pyproject.toml
[tool.poetry]
name = "techtest"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.8.2"
fastapi = "^0.100.0"
uvicorn = "^0.22.0"
sqlalchemy = "^2.0.0"
redis = "^4.0.0"
python-jose = "^3.3.0"
python-multipart = "^0.0.6"
httpx = "^0.24.0"
python-dotenv = "^0.21.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[tool.pytest.ini_options]
addopts = "-ra"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
